requirements.txt
# Core web framework
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
python-dotenv==0.19.2
requests==2.28.1

# File type detection
python-magic-bin==0.4.14  # Windows-compatible replacement

# Security analysis tools
bandit==1.7.4
flawfinder==2.0.19

# Git repository handling
gitpython==3.1.30

# ML & NLP dependencies
scikit-learn==1.0.2
joblib==1.1.1
numpy==1.22.4
pandas==1.4.3
nltk==3.8.1        # Bổ sung NLP xử lý văn bản
transformers==4.29.2  # Chuẩn bị sẵn cho BERT/Transformer nếu cần

# Frontend / Visualization
streamlit==1.11.0
altair>=4.0.0
matplotlib==3.5.2
seaborn==0.11.2

# Optional: Bảo mật và kiểm thử
pytest==7.1.2
flake8==5.0.4
